requirements.txt
torch==2.1.1+computecanada
torchvision==0.16.1+computecanada
scikit_learn==1.3.1+computecanada
pillow==9.4.0+computecanada
soundfile==0.12.1+computecanada